Add a new option to specify the zone name.
This is a default option which can be passed for both
master mode and node mode.

For the ovnkube-master, the passed zone name is validated
against the zone name stored in the Northbound database.
ovnkube-master will exit if there is a mismatch between
the two.  This ensures that ovnkube-master is connected
to the proper Northboubd database.

For the ovnkube-node, the passed zone name is validated against
the zone name stored in the Southbound databae.  ovnkube-node
will exit if there is a mismatch between the two.  This ensures that
ovnkube-node is connected to the proper Southbound database.

Signed-off-by: Numan Siddique <numans@ovn.org>