Add a new option to specify the zone name.
For the ovnkube-network-controller-manager, the passed zone name
is validated against the zone name stored in the Northbound database.
ovnkube-network-controller-manager  will exit if there is a
mismatch between the two.  This ensures that
ovnkube-network-controller-manager  is connected to the proper
Northboubd database.

For the ovnkube-node, the passed zone name is validated against
the zone name stored in the Southbound databae.  ovnkube-node
will exit if there is a mismatch between the two.  This ensures that
ovnkube-node is connected to the proper Southbound database.

Signed-off-by: Numan Siddique <numans@ovn.org>