EIP: Add support for compatibility with ETP=local
This commit introduces qos rules on every node's switch
to mark packets emerging from EIP pods if they are not
reply packets. These are the packets that need to be
reRouted to other node in the cluster. So basically we
mark the SYN packets that are matched on by the LRP we
saw in the previous commit. SYNACK packets are not marked
as these are responses to ingress/service traffic for
ETP=local.

sample:
from-lport   103 (ip4.src==$a4548040316634674295 && ct.trk && !ct.rpl) mark=42

This let's these two features to be used at the same time.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>