Handle upgrades
We need to update the LRPs for local pods such that
the new match takes effect. I could do this in two ways:

1) Keep the policy externalIDs as is and just query
for the local zone pod related policies and only update
their match; but there was no way to make this a "one time
only operation" because we don't have a good way to
differentiate this unless we compare the matches always
OR
2) Use this opportunity to start using new ExternalIDs
that way more meaningful - we might not be levering the
full dbIndexes here because other LRPs don't use them yet
BUT this can be starting ground to slowly convert them
one by one. So in this case I am actually nuking the old LRPs
without the externalIDs and then createReroutePolicyOps will
take of correctly creating the new ones.
Concerns:
1) maybe recreating all policies might cause load? but note
that local v/s remote pods need to be handled differently
wrt to match + we need to now add the new externalIDs to
all LRPs at 100. Adding a whole new syncer is beyond the scope
of this PR, so if nuking and recreating is bad I can try to
go with approach (1)

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>