Add masqueradeIP flows back for advertised networks in breth0
In the previous commits we added SNATing to nodeIP
for the following traffic flows:

pod -> nodes
pod -> nodeports

when pods are part of advertised networks. Prior to
SNATing to nodeIPs they are SNATed at the ovn_cluster_router
to masqueradeIP before being sent into the host.

In commit https://github.com/ovn-kubernetes/ovn-kubernetes/commit/75dd73fb645bff6c30e8c08c9b7f711d82996601
we had converted all UDN flows that matched on masqueradeIP
as the source on breth0 for UDN pods to services traffic flow
to instead match on the podsubnets.

However given we have pod to node and pod to nodeport
traffic flows using masqueradeIP as the SNAT we need to
now re-add the masqueradeIP flows as well to ensure that
nodeports isolation between UDNs work correctly.

Before this commit:

In LGW/SGW flow is: UDN pod -> samenodeIP:nodeport in default network ->
SNATed to masqueradeIP of that UDN -> sent to host -> SNATed to clusterIP ->
hits the default flow in table=2 in br-ex:

 cookie=0xdeff105, duration=15690.053s, table=2, n_packets=0, n_bytes=0, idle_age=15690, priority=100 actions=mod_dl_dst:6e:4d:97:c0:3c:97,output:2

and sends to patch port of default network and this traffic
starts working when it shouldn't.

In case of L3 UDN advertised pod -> nodeport service in default or other UDN network:
https://github.com/ovn-kubernetes/ovn-kubernetes/pull/4705/commits/d63887ed167da260d3f26c71ec06e520d89a4b0f
is the commit where we added logic to match on srcIP of the traffic and
accordingly route it into the respective UDN patchports. So there we use
the masqueradeIP of a particular UDN to determine what the source of the traffic
was and route it into that particular UDN's patchport where it would backhole
if there was no matching clusterIP NAT entry there, and this is how
isolation was guaranteed.

For l2 topology the logic is same as above for clusterIPs but
for nodeports the GR itself drops the packets destined
towards the other networks as there is no LB entry present on
the GR as the destination IP is that of the router itself. That's how
isolation works there:

sample trace:
    next;
10. ls_out_apply_port_sec (northd.c:6039): 1, priority 0, uuid 2aa6ebd5
    output;
    /* output to "stor-cluster_udn_tenant.blue.network_ovn_layer2_switch", type "l3gateway" */

ingress(dp="GR_cluster_udn_tenant.blue.network_ovn-worker2", inport="rtos-cluster_udn_tenant.blue.network_ovn_layer2_switch")
-----------------------------------------------------------------------------------------------------------------------------
 0. lr_in_admission (northd.c:13232): eth.dst == 0a:58:64:41:00:03 && inport == "rtos-cluster_udn_tenant.blue.network_ovn_layer2_switch", priority 50, uuid 7f9af183
    reg9[1] = check_pkt_larger(1414);
    xreg0[0..47] = 0a:58:64:41:00:03;
    next;
 1. lr_in_lookup_neighbor (northd.c:13420): 1, priority 0, uuid d2672052
    reg9[2] = 1;
    next;
 2. lr_in_learn_neighbor (northd.c:13430): reg9[2] == 1 || reg9[3] == 0, priority 100, uuid 84ca0ef4
    mac_cache_use;
    next;
 3. lr_in_ip_input (northd.c:12824): ip4.dst == {172.18.0.4}, priority 60, uuid ea41c4e7
    drop;

Without this fix:

[FAIL] BGP: isolation between advertised networks Layer3 connectivity between networks [It] pod in the UDN should not be able to access a default network service

the above test will work in LGW when it should not work like is
the case for non-advertised UDNs.

This commit adds back the masqueradeIP flow as well for advertised
networks and given that all traffic destined to any different network's
services would always be routed into its own patch port where it
gets blackholed, we guarantee isolation.

when 2 UDNs are advertised:

this PR added back these two flows with masqueradeIP match:
cookie=0xdeff105, duration=127.593s, table=2, n_packets=0, n_bytes=0, priority=200,ip,nw_src=169.254.0.12 actions=mod_dl_dst:de:9f:9a:36:10:74,output:3
cookie=0xdeff105, duration=127.534s, table=2, n_packets=0, n_bytes=0, priority=200,ip,nw_src=169.254.0.14 actions=mod_dl_dst:de:9f:9a:36:10:74,output:4

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>