UDN: Change priority of marked flows from 250 to 150
The traffic path for the pkt_mark flows in UDN is mentioned
here: https://github.com/ovn-kubernetes/ovn-kubernetes/commit/931facbe7161dc87b2eebafda32ca30714fd5dd3

it was again changed to be priority 250 with a drop action instead of reroute in this change:
https://github.com/ovn-kubernetes/ovn-kubernetes/pull/5351/commits/dcc403c1ddf11e30e6990699616405f6dc47dd71

It was mostly added for external clients to services support
in UDNs.

However right now I am hitting this issue where isolation is
broken between different UDNs i.e

Given these two flows match on same type of traffic:

 cookie=0xdeff105, duration=1951.037s, table=2, n_packets=2, n_bytes=148, idle_age=1913, priority=200,ip,nw_src=169.254.0.44 actions=drop
 cookie=0xdeff105, duration=1950.419s, table=2, n_packets=0, n_bytes=0, idle_age=1950, priority=200,ip,nw_src=169.254.0.46 actions=drop
 cookie=0xdeff105, duration=1951.037s, table=2, n_packets=0, n_bytes=0, idle_age=1951, priority=250,pkt_mark=0x1011,ip actions=mod_dl_dst:82:42:53:a3:9f:f4,output:19
 cookie=0xdeff105, duration=1951.037s, table=2, n_packets=0, n_bytes=0, idle_age=1951, priority=250,pkt_mark=0x1011,ipv6 actions=mod_dl_dst:82:42:53:a3:9f:f4,output:19
 cookie=0xdeff105, duration=1950.419s, table=2, n_packets=0, n_bytes=0, idle_age=1950, priority=250,pkt_mark=0x1012,ip actions=mod_dl_dst:82:42:53:a3:9f:f4,output:20
 cookie=0xdeff105, duration=1950.419s, table=2, n_packets=18, n_bytes=1872, idle_age=107, priority=250,pkt_mark=0x1012,ipv6 actions=mod_dl_dst:82:42:53:a3:9f:f4,output:20
 cookie=0xdeff105, duration=1951.037s, table=2, n_packets=0, n_bytes=0, idle_age=1951, priority=200,ipv6,ipv6_src=fd69::2c actions=drop
 cookie=0xdeff105, duration=1950.419s, table=2, n_packets=0, n_bytes=0, idle_age=1950, priority=200,ipv6,ipv6_src=fd69::2e actions=drop

based on the priority, the pkt_mark flows kick in
instead of src based flows, we break isolation since then the UDN pod
in network A can end up reaching the nodeport services in
network B because of nft rules marking the packets w/o
distinction coming from all UDNs.

This is particularly visible for advertised UDNs since based
on when the toggle for network is advertised the flows can be
inserted in any order. - After the PR that merged last week i.e
https://github.com/ovn-kubernetes/ovn-kubernetes/pull/5140
this bug is now present on master and there is no more ambiguity on
which flows kick in. It's always the pkt_mark flows.

I don't think we need to keep these flows in same priority?
Hence moving the marked flows to lower priority. -> REDUDANT question with 5140.
We need to keep pkt_mark flows at lower priority than the drop flows

Without this fix the following test will be broken:

Summarizing 1 Failure:
  [FAIL] BGP: isolation between advertised networks Layer2 connectivity between networks
[It] UDN pod to the same node nodeport service in different UDN network should not work

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>